Fix merge error + moved gtcallable back into the ffront package
